### 2023-06-11 #5539 @editorialbot
[PRE REVIEW]: mpl-interactions: A Python package for Interactive Matplotlib Figures
**Submitting author:** @ianhi (<a href="http://orcid.org/0000-0002-7591-083X">Ian Hunt-Isaak</a>)
**Repository:** https://github.com/mpl-extensions/mpl-interactions
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.23.0
