### 2023-07-05 #5539 @ianhi
Hi @ppxasjsm I'm sorry for the delay! I missed the initial message through a combination of attending a wedding and the subsequent holiday in the U.S. I'm now back and should be able to see messages much more promptly.

I spent some time searching the reviewer list and came up with the following as start. Happy to suggest more if need be.

flekschas - https://reviewers.joss.theoj.org/reviewers/2632
rgerum - https://reviewers.joss.theoj.org/reviewers/1336
paxtonfitzpatrick - https://reviewers.joss.theoj.org/reviewers/2369
arjunsavel - https://reviewers.joss.theoj.org/reviewers/1410
a-slide - https://reviewers.joss.theoj.org/reviewers/138
jkanche - https://reviewers.joss.theoj.org/reviewers/518
jonatanalvelid - https://reviewers.joss.theoj.org/reviewers/1803
desilinguist - https://reviewers.joss.theoj.org/reviewers/3
jochym - https://reviewers.joss.theoj.org/reviewers/54
rmhogervorst - https://reviewers.joss.theoj.org/reviewers/139
ranzhengcode - https://reviewers.joss.theoj.org/reviewers/2388
