### 2023-07-08 #5539 @ppxasjsm
Hi @flekschas, would you be willing to review this JOSS submission?
