### 2023-07-13 #5651 @rgerum
The authors provide an extension for matplotlib to add interactivity to plots by providing parametersliders. While this is not exactly a new idea as both matplotlib and ipyhon already feature sider widgets, the authors point out that these solutions are not handling the plot update in a performant maner and are also rather difficult to use.
The API seems well tought through and offers different options on how to add sliders, have joined sliders to inlucence multiple plots, so I think the package has a substantial size and therefore meets the scope of JOSS. The package has automatic test, although they are currently (according to Github) failing for mpl-latest, but for the stable versions of matplotlib they seem to run well.

The only checkmark I haven't set yet is about the list of authors (Contribution and authorship). There are a good number of other contributors listed in the github repository readme that do not appear as authors of the paper. Also Doeke Hekstra does not appear in the contributor list in the reposiotires readme. Could the author elaborate on the reasons for this?
