### 2023-07-18 #5651 @flekschas
**General comment**

`mpl-interactions` is an extension for `matplotlib` for automating the creation of interactive sliders, inputs, radio buttons for configuring a visualization. While the extension does not seem to offer general interactions/interactivity (like view manipulations or selections), being able to rapidly adjust plotting and function parameters could be to be useful to help scientist to better understand their data and methods.
The paper is well written and contains easy to understand examples. The software itself is extensively document and seems to be in a well-maintained state.

Below I list why I cannot yet check every box:

**Contribution and authorship**

I agree with @rgerum here. Several contributors with substantial contributions (e.g., with more than 10 commits) are not listed as an author but Doeke Hekstra with zero code contributions is listed as an author. Clarification is needed.

**State of the field:**

The authors have commented on the shortcomings of interactive figures with `matplotlib` and `ipywidgets` but they left out the shortcomings of `mpl-interactions`: `ipywidgets` can easily be integrated with other widgets while from the documentation and code samples, it looks like `mpl-interactions` can really only be used to interactively plot matplotlib figures faster. Moreover, the extension does not offer view manipulations like pan&zoom or selections. It'd be great to clarify this. Also, why is `ipympl` not mentioned?

And finally, on a broader level, one might also wonder how `mpl-interactions` compares to other interactive plotting libraries like altair, holoviz, plotly, etc. Providing a broader perspective would be really useful to the reader.

**Installation & Functionality**

I followed https://mpl-interactions.readthedocs.io/en/stable/install.html#setup-for-jupyterlab-3 and tried to run the _Basic Example_ from https://mpl-interactions.readthedocs.io/en/stable/index.html#basic-example with JupyterLab v4.0.2 on a M1 MacBook but can't get it to work.

Not sure what I might have done wrong or missing, but unfortunately, I didn't get `mpl-interactions` to work.

**Minor writing comments**

> the choices it makes are not always optimal for scientific plotting

This statement is fairly strong but unjustified. The authors should either give a definition + example of what optimality for scientific plotting  means and why ipywidgets does not support it. Or alternatively remove this statement.

> `y_data = logistic_growth(t_data, L=5, k=1, t0=1) + rng.normal(size=t_data.size, scale=0.`

This line is cut off

> This framework makes it easy generate complex interactive visualizations

A `to` is missing after `easy`

> Complete Tutorials, Examples, and API documentation are available on https://mpl-interactions.readthedocs.io/en/stable/

Why are `Tutorials` and `Examples` examples capitalized here?
