### 2023-07-21 mpl-extensions/mpl-interactions#282 @ianhi
joss review: writing comments
From @flekschas in https://github.com/openjournals/joss-reviews/issues/5651#issuecomment-1641072224

>     the choices it makes are not always optimal for scientific plotting
>
> This statement is fairly strong but unjustified. The authors should either give a definition + example of what optimality for scientific plotting means and why ipywidgets does not support it. Or alternatively remove this statement.

I definitely agree that this is currently unjustified in the paper. The differences referred to are laid out here: https://mpl-interactions.readthedocs.io/en/stable/comparison.html#differences-in-generated-widgets  Would you find these differences in generated widgets to be convincingly better for interactively plotting scientific models?

>     y_data = logistic_growth(t_data, L=5, k=1, t0=1) + rng.normal(size=t_data.size, scale=0.
>
> This line is cut off

good catch, will reformat the code.

>     This framework makes it easy generate complex interactive visualizations
>
> A to is missing after easy

thank you.

>     Complete Tutorials, Examples, and API documentation are available on https://mpl-interactions.readthedocs.io/en/stable/
>
> Why are Tutorials and Examples examples capitalized here?

will correct
