### 2023-07-21 mpl-extensions/mpl-interactions#283 @ianhi
Joss review: authorship
from @rgerum in: https://github.com/openjournals/joss-reviews/issues/5651#issuecomment-1634353639

> The only checkmark I haven't set yet is about the list of authors (Contribution and authorship). There are a good number of other contributors listed in the github repository readme that do not appear as authors of the paper. Also Doeke Hekstra does not appear in the contributor list in the reposiotires readme. Could the author elaborate on the reasons for this?

from @flekschar:

> I agree with @rgerum here. Several contributors with substantial contributions (e.g., with more than 10 commits) are not listed as an author but Doeke Hekstra with zero code contributions is listed as an author. Clarification is needed.

> (e.g., with more than 10 commits)

A small note re commits is that these are not a perfect measure as at some point in the repositories history I switched to squash merging from merging the branch with all commits.

**Doeke Hekstra**

Doeke (@DHekstra) has not contributed code to this project, however he was integral to the genesis of the project. We were analyzing data together, and our discussions of how best to visualize the data and our models are what led to this package. To us this felt like a significant intellectual contribution.

**Other contributors**

There are many contributors of small fixes, features, or bug reporters. I've done my best to encourage this and also acknowledge these people in the github readme. Regarding authorship, we weren't sure quite what the bounds of appropriate authorship were and if they could include a focus on single features and or documentation vs overall structure of the package. I definitely prefer to be more inclusive than exclusive in authorship. I would propose adding @jrussell25, @redeboer and @samanthahamilton as authors who put in sustained effort to various aspects of the package.
