### 2023-07-27 mpl-extensions/mpl-interactions#283 @rgerum
as you already have a section in the readme to thank everyone who did some contributions it would be good to add Doeke Hekstra there if he was an integral part of the project.

If these other authors have put in enough effort according to you, they should be authors of the paper I think. I will change my review accordingly when this is implemented.
