### 2023-07-27 mpl-extensions/mpl-interactions#283 @flekschas
I agree with @rgerum.
