### 2023-07-27 mpl-extensions/mpl-interactions#282 @flekschas
> I definitely agree that this is currently unjustified in the paper. The differences referred to are laid out here: https://mpl-interactions.readthedocs.io/en/stable/comparison.html#differences-in-generated-widgets Would you find these differences in generated widgets to be convincingly better for interactively plotting scientific models?

Even though this is software-focused paper, the paper still needs to justify statements. You could link to the comparison but to be honest I'm not convinced that the comparison justifies the statement about optimality. I would just focus on the three key features that are laid out in the comparison: improved re-rendering performance, portability, and convenience (for when frequent re-rendering is needed). Those are concrete features for which you have convincing arguments that mpl-interactions improves over ipywidgets. Whether those three features define optimality for scientific plotting is a different question that you could only answer by conducting a user study or survey.
