### 2023-07-28 mpl-extensions/mpl-interactions#281 @flekschas
Create a fresh environment from scratch works! 🎉 Thanks for the detailed description/response. Will update my review accordingly
