### 2023-07-28 mpl-extensions/mpl-interactions#281 @flekschas
Feel free to close this issue :)
