### 2023-08-08 mpl-extensions/mpl-interactions#283 @ianhi
bump to @jrussell25 @redeboer @samanthahamilton re authorship of the joss paper. If you are interested in being added to the author list please try to let me know by Monday August 14. Happy to talk offline (as it were) as well - you can email me `ianhuntisask` at harvard.edu
