### 2023-08-08 mpl-extensions/mpl-interactions#283 @jrussell25
I'd be very happy to be included as an author!
