### 2023-08-08 mpl-extensions/mpl-interactions#284 @ianhi
[JOSS review]: state of the field
From @flekschas in https://github.com/openjournals/joss-reviews/issues/5651#issuecomment-1641072224

> State of the field:
> The authors have commented on the shortcomings of interactive figures with matplotlib and ipywidgets but they left out the shortcomings of mpl-interactions:  ipywidgets can easily be integrated with other widgets while from the documentation and code samples, it looks like mpl-interactions can really only be used to interactively plot matplotlib figures faster.

This core purpose of mpl-interactions is indeed solely to make it easier to create plots in matplotlib that respond to changes in widgets. However, the generated widgets (either `matplotlib.widgets` or `ipywidgets`) are not limited to this. They can be used as normal widgets an integrated into any other widget elsewhere in user code. There is also a mechanism to connect arbitrary functions as callbacks to the generated widgets (https://mpl-interactions.readthedocs.io/en/stable/examples/custom-callbacks.html) which allows coupling the generated widgets to other components of code. If a user wants to use an existing widget to also control the plot they can pass that widget into the `iplt` function https://mpl-interactions.readthedocs.io/en/stable/examples/usage.html#existing-widgets.

So in a general sense it is a limitation that mpl-interactions it is not designed to generate controls for arbitrary non-matplotlib functions. However, within the bounds of the problem space it sets out to solve I don't think it has clear drawbacks compared to a user generating their own widgets and manually controlling matplotlib. With the exception that it does not provide complete coverage of all matplotlib functions. Would it help to more clearly delineate problem to be solved and then focus on the relative drawbacks?

> Moreover, the extension does not offer view manipulations like pan&zoom or selections. It'd be great to clarify this.

pan/zoom are available through matplotlib via the figure toolbar

![image](https://github.com/mpl-extensions/mpl-interactions/assets/10111092/88f7aec3-89c6-40af-be0a-8ee86ec5262a)

originally `mpl-interactions` provided utilities for making these interactions more userfriendly: https://mpl-interactions.readthedocs.io/en/stable/examples/zoom-factory.html however, these are being slowly deprecated and moved to a standalone package [mpl-pan-zoom](https://mpl-pan-zoom.readthedocs.io/en/latest/) to allow mpl-interactions to focus on it's core purpose, generating widgets to control plotted elements.

>  or selections.

I'm not sure exactly what you mean here. Do you mean the ability to select elements within the plot using the mouse? This is not a feature that matplotlib provides and outside the core mission of this library. There is also an existing library [mpl-cursors](https://mplcursors.readthedocs.io/en/stable/) that provides what I suspect you are asking after.

I'd consider  the issues of zoom/pan/selection to be out of scope for this package as it's focus to facilitate changign the parameters of functions and replotting them. Perhaps the name of the package misleading? In that `interactions` could be taken to mean all types of interactivity. The name was originally meant to invoke the `ipywidgets.Interact` function, but to be more specific to matplotlib. However, I can see how it could lead to expectations of providing more types of interaction.

> Also, why is ipympl not mentioned?

ipympl was not mentioned as it is simply one of the interactive backends(https://matplotlib.org/stable/users/explain/backends.html#interactive-backends)  for maptlotlib. `mpl-interactions` will work with any of these backends, ipympl is simply the most natural one to use when working in a jupyter notebook. I can see how with the discussion of ipywidgets there may be confusion that ipympl is necessary.

> And finally, on a broader level, one might also wonder how mpl-interactions compares to other interactive plotting libraries like altair, holoviz, plotly, etc. Providing a broader perspective would be really useful to the reader.

We will do some experimenting with these other libraries as we have not used them extensively in our own work. Some of this may come down to differences in those library vs matplotlib. In a way `mpl-interactions` adds a feature to maptlotlib that some of those libraries have natively.

We will also try to include a comparison to panel, which is probably the most similar to this library in that it can generate widgets to control maptlotlib plots.

----

Sorry I took awhile to post this. I thought I caught this in my initial round of issue creation in response, though I clearly missed it.
