### 2023-08-18 mpl-extensions/mpl-interactions#284 @flekschas
Thank you for sharing your perspective. This all makes sense now.

> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

In order for me to check this box, I'd be great if you could include some of your thoughts from here into the paper to give the reader a better understanding of how `mpl-interactions` compares to the many other interactive data visualization tools. This doesn't have to be comprehensive but it'd be nice to know (as a reader) what `mpl-interactions` could be used for and what is outside the scope of `mpl-interactions`.

> I'm not sure exactly what you mean here. Do you mean the ability to select elements within the plot using the mouse? This is not a feature that matplotlib provides and outside the core mission of this library. There is also an existing library [mpl-cursors](https://mplcursors.readthedocs.io/en/stable/) that provides what I suspect you are asking after.

Yes. Selections are a key aspect of interactive data visualizations. E.g., select a set of discrete data points in a scatter plot or a range in a continuous function. See for instance https://vega.github.io/vega-lite/docs/selection.html. It'd be very useful for the reader to know that selections are outside the scope of `mpl-interactions`.

> I'd consider the issues of zoom/pan/selection to be out of scope for this package as it's focus to facilitate changign the parameters of functions and replotting them. Perhaps the name of the package misleading? In that interactions could be taken to mean all types of interactivity. The name was originally meant to invoke the ipywidgets.Interact function, but to be more specific to matplotlib. However, I can see how it could lead to expectations of providing more types of interaction.

This explain my confusion. Maybe you can add a sentence to the statement of need to clarify how you define interactions and what aspects you think are outside the scope of `mpl-interactions` but are part of the larger concept of interactive data visualizations.
