### 2023-08-31 mpl-extensions/mpl-interactions#282 @ianhi
This should all be resolved by https://github.com/mpl-extensions/mpl-interactions/pull/260/commits/e79a7b520fbb92a27a6f8d1542453f4007047d9a feel free to re-open if not
