### 2023-09-05 52a9ba0
paper: update statement of need for state of the field
paper: clarify out of scope

paper: reword

spelling: typo
