### 2023-09-05 3e83d9b
test: update tests for newer mpl (#287)
* test: update tests for newer mpl

* test: update test versions

* fix: typo

* update dep notice

* fix: more typos
