+++ .github/workflows/mpl-master-test.yml
@@ -17,7 +17,7 @@ jobs:
         python-version: ["3.x"]
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
