+++ .github/workflows/linkcheck.yml
@@ -15,7 +15,7 @@ jobs:
     name: Check external links
     runs-on: ubuntu-20.04
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Set up Python 3.9
         uses: actions/setup-python@v4
         with:
