+++ .github/workflows/test.yml
@@ -19,7 +19,7 @@ jobs:
         matplotlib-version: ["3.6", "3.7"]
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
