+++ README.md
@@ -9,9 +9,9 @@
 [![PyPI package](https://badge.fury.io/py/mpl-interactions.svg)](https://pypi.org/project/mpl-interactions)
 [![Supported Python versions](https://img.shields.io/pypi/pyversions/mpl-interactions)](https://pypi.org/project/mpl-interactions)
 [![Documentation Status](https://readthedocs.org/projects/mpl-interactions/badge/?version=stable)](https://mpl-interactions.readthedocs.io/en/stable/?badge=stable)
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/ianhi/mpl-interactions/master?urlpath=lab)
-[![Test](https://github.com/ianhi/mpl-interactions/actions/workflows/test.yml/badge.svg)](https://github.com/ianhi/mpl-interactions/actions/workflows/test.yml)
-[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/ianhi/mpl-interactions/master.svg)](https://results.pre-commit.ci/latest/github/ianhi/mpl-interactions/main)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/mpl-extensions/mpl-interactions/master?urlpath=lab)
+[![Test](https://github.com/mpl-extensions/mpl-interactions/actions/workflows/test.yml/badge.svg)](https://github.com/mpl-extensions/mpl-interactions/actions/workflows/test.yml)
+[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/mpl-extensions/mpl-interactions/master.svg)](https://results.pre-commit.ci/latest/github/mpl-extensions/mpl-interactions/main)
 [![Code style: Prettier](https://img.shields.io/badge/code_style-prettier-ff69b4.svg?style=flat-square)](https://prettier.io)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Imports: isort](https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336)](https://pycqa.github.io/isort)
@@ -20,7 +20,7 @@
 
 mpl_interactions' library provides helpful ways to interact with [Matplotlib](https://matplotlib.org/) plots. Full narrative documentation and example can be found on [Read the Docs](https://mpl-interactions.readthedocs.io).
 
-<img src=https://raw.githubusercontent.com/ianhi/mpl-interactions/master/docs/_static/images/short-interactive.gif width=45%> <img src=https://raw.githubusercontent.com/ianhi/mpl-interactions/master/docs/_static/images/heatmap_slicer.gif width=45%>
+<img src=https://raw.githubusercontent.com/mpl-extensions/mpl-interactions/master/docs/_static/images/short-interactive.gif width=45%> <img src=https://raw.githubusercontent.com/mpl-extensions/mpl-interactions/master/docs/_static/images/heatmap_slicer.gif width=45%>
 
 ## Installation
 
@@ -35,7 +35,7 @@ If you use JupyterLab, make sure you follow the full instructions in the ipympl 
 
 ## Contributing / feature requests / roadmap
 
-I use the GitHub [issues](https://github.com/ianhi/mpl-interactions/issues) to keep track of ideas I have, so looking through those should serve as a roadmap of sorts. For the most part I add to the library when I create a function that is useful for the science I am doing. If you create something that seems useful a PR would be most welcome so we can share it easily with more people. I'm also open to feature requests if you have an idea.
[…]
