### 2023-09-08 mpl-extensions/mpl-interactions#283 @ianhi
I've included John as an author https://github.com/mpl-extensions/mpl-interactions/pull/260/commits/fe8f16908248cc4b44424f4431a0520605b09d9d

and from my perspective both @samanthahamilton and @redeboer are both welcome to change their minds at any point until the paper is published.
