### 2023-09-08 mpl-extensions/mpl-interactions#284 @ianhi
We've attempted to address your comments with the changes in this commit: https://github.com/mpl-extensions/mpl-interactions/pull/260/commits/52a9ba0c6a6a5964f3f0fd579b58953f2d063f3e
