### 2023-09-20 mpl-extensions/mpl-interactions#284 @flekschas
Looks good but citations for Altair, Holoviz, and Plotly are missing. There's no harm in citing data visualization fellow packages. :)
