### 2023-09-20 #5651 @flekschas
I checked off "Contribution and authorship" as well for the same reason. I also checked "State of the field" given https://github.com/mpl-extensions/mpl-interactions/commit/52a9ba0c6a6a5964f3f0fd579b58953f2d063f3e#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424.

The last missing check for me is "References". Once the authors cite "Altair, Holoviz, and plotly" things are good to go on my end.
