+++ paper/paper.bib
@@ -13,9 +13,56 @@
   year      = 2007
 }
 
-@unpublished{interactive_Jupyter_widgets,
+@unpublished{interactive_Jupyter_widgets:2015,
 author = {Jupyter widgets community},
 title = {ipywidgets, a GitHub repository},
 year = {2015},
 note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets}
+}
+
+@article{VanderPlas2018, 
+   doi = {10.21105/joss.01057},
+   url = {https://doi.org/10.21105/joss.01057}, 
+   year = {2018},
+   publisher = {The Open Journal}, 
+   volume = {3}, 
+   number = {32},
+   pages = {1057},
+   author = {Jacob VanderPlas and Brian E. Granger and Jeffrey Heer and Dominik Moritz and Kanit Wongsuphasawat and Arvind Satyanarayan and Eitan Lees and Ilia Timofeev and Ben Welsh and Scott Sievert},
+   title = {Altair: Interactive Statistical Visualizations for Python}, 
+   journal = {Journal of Open Source Software} 
+} 
+
+@software{philipp_rudiger_2020_3634720,
+  author       = {Philipp Rudiger and
+                  Julia Signell and
+                  James A. Bednar and
+                  Andrew and
+                  Jean-Luc Stevens and
+                  Chris B and
+                  Jordan Samuels and
+                  Todd and
+                  Thomas PEDOT and
+                  Sander van den Oord and
+                  Jon Mease and
+                  Isaac Virshup and
+                  Gabriel Corona and
+                  Danny Hermes and
+                  Curtis H. and
+                  Anita Graser},
+  title        = {holoviz/hvplot: Version 0.5.2},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.5.2},
+  doi          = {10.5281/zenodo.3634720},
+  url          = {https://doi.org/10.5281/zenodo.3634720}
+}
+@online{plotly, 
+   author = {Plotly Technologies Inc.}, 
+   title = {Collaborative data science},
+   publisher = {Plotly Technologies Inc.},
+   address = {Montreal, QC}, 
+   year = {2015}, 
+   url = {https://plot.ly}
 }
