+++ paper/paper.bib
@@ -20,18 +20,18 @@ year = {2015},
 note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets}
 }
 
-@article{VanderPlas2018, 
+@article{VanderPlas2018,
    doi = {10.21105/joss.01057},
-   url = {https://doi.org/10.21105/joss.01057}, 
+   url = {https://doi.org/10.21105/joss.01057},
    year = {2018},
-   publisher = {The Open Journal}, 
-   volume = {3}, 
+   publisher = {The Open Journal},
+   volume = {3},
    number = {32},
    pages = {1057},
    author = {Jacob VanderPlas and Brian E. Granger and Jeffrey Heer and Dominik Moritz and Kanit Wongsuphasawat and Arvind Satyanarayan and Eitan Lees and Ilia Timofeev and Ben Welsh and Scott Sievert},
-   title = {Altair: Interactive Statistical Visualizations for Python}, 
-   journal = {Journal of Open Source Software} 
-} 
+   title = {Altair: Interactive Statistical Visualizations for Python},
+   journal = {Journal of Open Source Software}
+}
 
 @software{philipp_rudiger_2020_3634720,
   author       = {Philipp Rudiger and
@@ -58,11 +58,11 @@ note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets}
   doi          = {10.5281/zenodo.3634720},
   url          = {https://doi.org/10.5281/zenodo.3634720}
 }
-@online{plotly, 
-   author = {Plotly Technologies Inc.}, 
+@online{plotly,
+   author = {Plotly Technologies Inc.},
    title = {Collaborative data science},
    publisher = {Plotly Technologies Inc.},
-   address = {Montreal, QC}, 
-   year = {2015}, 
+   address = {Montreal, QC},
+   year = {2015},
    url = {https://plot.ly}
 }
