### 2023-09-20 mpl-extensions/mpl-interactions#284 @ianhi
definitely! I remember thinking I should add them but it must've slipped through thanks for the catch.

Added in https://github.com/mpl-extensions/mpl-interactions/pull/260/commits/8cbc9f79053b22a35c7ed4d109971cf198cf8579
