### 2023-09-20 #5651 @ianhi
> The package has automatic test, although they are currently (according to Github) failing for mpl-latest, but for the stable versions of matplotlib they seem to run well.

I forgot to mention but I fixed this a week or sot ago, tests now pass on the master branch of matplotlib.
