### 2023-09-20 mpl-extensions/mpl-interactions#284 @flekschas
Awesome thanks! Will check the last remaining checkbox
