### 2023-10-04 #5651 @ppxasjsm
@ianhi here are some editorial suggestions for the paper:

- [ ] capitalize the word package in the title
- [ ] "matplotlib" should be consistently capitalized as "Matplotlib".
- [ ]  "mpl-interactions" should be consistently formatted with hyphens
- [ ] In the sentence "Matplotlib provides mechanisms for updating elements (artists) in figures," there's a missing comma before "artists." It should be: "Matplotlib provides mechanisms for updating elements (artists) in figures."
- [ ] overlaping -> overlapping
- [ ] "Thus it enhances an analysis workflow that [...]" -> "Thus, it enhances an analysis workflow that" Also maybe break this sentence into two. It is long.
- [ ] "Then mpl-interactions will generate the appropriate [...]" -> "Then, mpl-interactions will generate the appropriate"
