### 2023-10-04 #5651 @ppxasjsm
The ipywidgets reference appears a bit weird. I am not sure how to suggest an improvement but maybe making sure that a link is included would be a good first step. The editor-in-chief who will process the final version may have a suggestion here.
