### 2023-10-05 8afe522
citations: improve ipywidgets citation
