+++ paper/paper.bib
@@ -15,9 +15,9 @@
 
 @unpublished{interactive_Jupyter_widgets:2015,
 author = {Jupyter widgets community},
-title = {ipywidgets, a GitHub repository},
+title = {Ipywidgets},
 year = {2015},
-note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets}
+url = {https://github.com/jupyter-widgets/ipywidgets}
 }
 
 @article{VanderPlas2018,
