### 2023-10-05 469cc6b
paper: consistent matplotlib capitalization
