+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "mpl-interactions: A Python package for Interactive Matplotlib Figures"
+title: "mpl-interactions: A Python Package for Interactive Matplotlib Figures"
 tags:
   - Python
   - Matplotlib
@@ -41,11 +41,11 @@ The ability to interact dynamically with plots through widgets such as sliders c
 
 Matplotlib provides mechanisms for updating elements (artists) in figures. However, the APIs for these artists are not consistent and some are under- or undocumented. Furthermore, the creation and positioning of the native Matplotlib widgets is nontrivial. While the `ipywidgets` [@interactive_Jupyter_widgets:2015] library makes widget creation and positioning easier, it is difficult to integrate with Matplotlib in a performant manner. The easiest way to do so is to use the `ipywidgets`' `interact()` function, which automatically generates sliders and other widgets to control arguments to arbitrary python functions. However, this can result in completely regenerating the figure which can be slow. Alternatively, the user needs to remember the specifics of how to update each individual artist. While `Matplotlib` and `ipywidgets` provide the tools for controlling plots with widgets, the overhead of implementing such control can overwhelm its utility. `mpl-interactions` fills this gap by making it easy for users to generate widgets that dynamically control plots.
 
-There are a wide range of data visualization tools for Python, such as Altair [@VanderPlas2018], Holoviz[@philipp_rudiger_2020_3634720], and Plotly[@plotly], which provide rich interactive plotting experiences. These tools often have overlaping functionality with `mpl-interactions` and in some cases provide a greater range of interactive capabilities. However, they may not be useful for a user already invested in using Matplotlib, or for a user for whom Matplotlib is otherwise the best solution. For such users `mpl-interactions` adds functionality to `Matplotlib`. To this end `mpl-interactions` closely follows the semantics of the Matplotlib API, rather than creating a separate set of semantics, like the other mentioned libraries. Thus it enhances an analysis workflow that uses Matplotlib by enabling users to add interactive features to a library they are already using, rather than using different libraries for different aspects of the data analysis process.
+There are a wide range of data visualization tools for Python, such as Altair [@VanderPlas2018], Holoviz[@philipp_rudiger_2020_3634720], and Plotly[@plotly], which provide rich interactive plotting experiences. These tools often have overlapping functionality with `mpl-interactions` and in some cases provide a greater range of interactive capabilities. However, they may not be useful for a user already invested in using Matplotlib, or for a user for whom Matplotlib is otherwise the best solution. For such users `mpl-interactions` adds functionality to `Matplotlib`. To this end `mpl-interactions` closely follows the semantics of the Matplotlib API, rather than creating a separate set of semantics, like the other mentioned libraries. Thus, it enhances an analysis workflow that uses Matplotlib by enabling users to add interactive features to a library they are already using. Otherwise user would need to use multiple plotting libraries for different aspects of the data analysis process.
 
 # Overview
 
-`mpl-interactions` provides several key features to make generating interactive figures simple. The first is what arguments are accepted. While `Matplotlib` requires users to pass arrays as arguments, `mpl-interactions` allows passing a function that returns numeric values. Parameters to these functions are specified by adding extra keyword arguments (`kwargs`) to the plotting function call. Then `mpl-interactions` will generate the appropriate widgets for the parameters and run the functions to generate the numerical data to plot. For example, to plot a sinusoid and control its amplitude and frequency using sliders, a function returning the `y` values is defined and passed as the `y` parameter to the `plot` function. The ranges of the `A` and `f` parameters are defined as extra keyword arguments using tuples as a shorthand for what widget to generate.
[…]
