### 2023-10-05 #5651 @ianhi
- [x] capitalize the word package in the title
- [x] "matplotlib" should be consistently capitalized as "Matplotlib".
    - done, excluding code snippets
- [x] overlaping -> overlapping
- [x] "Thus it enhances an analysis workflow that [...]" -> "Thus, it enhances an analysis workflow that" Also maybe break this sentence into two. It is long.
- [x] "Then mpl-interactions will generate the appropriate [...]" -> "Then, mpl-interactions will generate the appropriate"

>In the sentence "Matplotlib provides mechanisms for updating elements (artists) in figures," there's a missing comma before "artists." It should be: "Matplotlib provides mechanisms for updating elements (artists) in figures."

I'm a little confused, "(artists)" is providing an alterntaive name for "elements" so the comma should come after if anything?

> "mpl-interactions" should be consistently formatted with hyphens

I've made sure that this is the case true everywhere except in code snippets.

> The ipywidgets reference appears a bit weird. I am not sure how to suggest an improvement but maybe making sure that a link is included would be a good first step. The editor-in-chief who will process the final version may have a suggestion here.

There is unfortunately no guidance from the project for citing. So originally the citation was based on https://discourse.jupyter.org/t/how-to-cite-jupyter-interactive-widgets-in-an-academic-paper/4435 which as you noted does render well with this template. So I've updated it to the following:

![image](https://github.com/openjournals/joss-reviews/assets/10111092/4ea526d0-0ebc-4b08-b86e-c195cd2287fb)

---
