### 2023-10-09 #5651 @ppxasjsm
Ignore the artist comment, not sure where my head was at with this.

Thank you for implementing all the other changes!

If you can now prepare a release and post it on zenodo and then follow the other tasks that would be amazing.
