+++ paper/paper.bib
@@ -14,7 +14,7 @@
 }
 
 @unpublished{interactive_Jupyter_widgets:2015,
-author = {Jupyter widgets community},
+author = {Community},
 title = {Ipywidgets},
 year = {2015},
 url = {https://github.com/jupyter-widgets/ipywidgets}
