### 2023-12-07 #5651 @ianhi
Hi @ppxasjsm I'm very sorry about the delays. I am now officially back to being able to keep track of conversations on github.

- [x] Double check
- [x] release: `0.24.1`
- [x] zenodo: https://doi.org/10.5281/zenodo.10211398
- [x]  title + author list match
- [x] license: both are 3 clause bsd
