### 2024-01-31 #5651 @arfon
@ianhi – I made a couple of tiny changes to formatting here: https://github.com/mpl-extensions/mpl-interactions/pull/290

Could you please merge this and I can proceed to accept?
