### 2024-01-31 #5651 @arfon
@flekschas, @rgerum - many thanks for your reviews here and to @ppxasjsm for editing this submission.

@ianhi - your paper is now accepted and published in JOSS ⚡️🚀💥
